Fix Windows x64 code generation for 64-bit types and mixed parameters
- virtual.c: Use sizeof(uintptr_t) instead of sizeof(long) in
  is_convert_noop() to correctly handle platforms where long is 32-bit
  but pointers are 64-bit (Windows x64)

- x86_64.c: Add proper unsigned-to-float conversions for DILL_U and
  DILL_US types by zero-extending to 64-bit before cvtsi2ss/cvtsi2sd

- x86_64_rt.c: Fix DtoUL helper to cast directly to size_t instead of
  going through long (which truncates on Windows where long is 32-bit)

- vtests/mixed_params.c: Add test for mixed integer/float parameter
  passing which exercises the Windows x64 calling convention

+++ virtual.c
@@ -3881,7 +3881,8 @@ is_convert_noop(int insn_code)
     int to_type = insn_code & 0xf;
 
     /* GSE -bug  This test should be for *generated* target, not host */
-    if (sizeof(long) != sizeof(int)) {
+    /* Use sizeof(uintptr_t) since that's what DILL uses for UL/L types */
+    if (sizeof(uintptr_t) != sizeof(int)) {
         return 0;
     } else {
         switch (from_type) {
